Feat: some Advancements were moved to their from the author intended level
