+++ src/main/kotlin/live/einfachgustaf/mods/smp/advancement/impl/challenging/NotSilkyEnough.kt
@@ -1,4 +1,4 @@
-package live.einfachgustaf.mods.smp.advancement.impl.normal
+package live.einfachgustaf.mods.smp.advancement.impl.challenging
 
 import kotlinx.coroutines.launch
 import live.einfachgustaf.mods.smp.advancement.Advancements
